refactor(packaging): viewmodel
