+++ app/src/main/java/com/example/ezvault/viewmodel/SelectedItemViewModel.java
@@ -1,4 +1,4 @@
-package com.example.ezvault;
+package com.example.ezvault.viewmodel;
 
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
@@ -8,7 +8,7 @@ import com.example.ezvault.model.Item;
 /**
  *
  */
-public class ItemViewModel extends ViewModel {
+public class SelectedItemViewModel extends ViewModel {
     private final MutableLiveData<Item> itemData = new MutableLiveData<>();
 
     public void set(Item item) {
